Grandtab

GrandTab data is a compilation of sources estimating the late-fall, winter, spring, and fall-run Chinook salmon populations for streams surveyed. Estimates are provided by the California Department of Fish and Wildlife, the US Fish and Wildlife Service, the California Department of Water Resources, the East Bay Municipal Utilities District, the US Bureau of Reclamation, the Lower Yuba River Management Team, and the Fisheries Foundation of California. For more information see CDFW Anadromous Assessment

grandtab_observed

grandtab_imputed

Observed

This data will be used in the fitness function to measure the difference between model predictions and observed escapement. Missing values are NA and we made all records less than 100 NA to account for a lack of confidence for counts less than 100. For spring run, counts less than 100 are included. Feather and Yuba Rivers do not separate spring and fall run, these estimate were scaled using a mean proportion for the runs from 2010-2012 coded wire tag data.

Imputed

This data will be used to calculate the number of juveniles during the 20 year simulation. The GrandTab data is incomplete for many watersheds during the 20 year period of calibration. For watersheds with no GrandTab data, we used 40 as the default escapement value. For watersheds with incomplete data, we used the mean escapement value.